Add dedicated factories for DecryptionProfile models
Move DecryptionProfile-related factories to a new module for better organization and maintainability. Update imports and tests accordingly to reflect the new module structure.
