+++ tests/factories.py
@@ -21,22 +21,12 @@ from scm.models.network.nat_rules import (
 
 # Local SDK imports
 from scm.models.security import (
-    DecryptionProfileCreateModel,
-    DecryptionProfileResponseModel,
     DNSSecurityProfileCreateModel,
     DNSSecurityProfileResponseModel,
     SecurityRuleCreateModel,
     SecurityRuleMoveModel,
     VulnerabilityProfileCreateModel,
     VulnerabilityProfileResponseModel,
-)
-from scm.models.security.decryption_profiles import (
-    DecryptionProfileUpdateModel,
-    SSLForwardProxy,
-    SSLInboundProxy,
-    SSLNoProxy,
-    SSLProtocolSettings,
-    SSLVersion,
 )
 from scm.models.security.dns_security_profiles import (
     ActionEnum,
@@ -82,311 +72,6 @@ from scm.models.security.wildfire_antivirus_profiles import (
     WildfireAvRuleBase,
     WildfireAvThreatExceptionEntry,
 )
-
-# ----------------------------------------------------------------------------
-# Decryption Profile object factories.
-# ----------------------------------------------------------------------------
-
-
-# Sub factories
-class SSLProtocolSettingsFactory(factory.Factory):
-    """Factory for creating SSLProtocolSettings instances."""
-
-    class Meta:
-        model = SSLProtocolSettings
-
-    auth_algo_md5 = True
-    auth_algo_sha1 = True
-    auth_algo_sha256 = True
-    auth_algo_sha384 = True
-    enc_algo_3des = True
-    enc_algo_aes_128_cbc = True
-    enc_algo_aes_128_gcm = True
-    enc_algo_aes_256_cbc = True
-    enc_algo_aes_256_gcm = True
-    enc_algo_chacha20_poly1305 = True
-    enc_algo_rc4 = True
-    keyxchg_algo_dhe = True
-    keyxchg_algo_ecdhe = True
-    keyxchg_algo_rsa = True
-    max_version = SSLVersion.tls1_2
-    min_version = SSLVersion.tls1_0
-
-    @classmethod
-    def with_versions(cls, min_ver: SSLVersion, max_ver: SSLVersion, **kwargs):
-        """Create an instance with specific SSL/TLS versions."""
-        return cls(min_version=min_ver, max_version=max_ver, **kwargs)
-
-
-class SSLForwardProxyFactory(factory.Factory):
-    """Factory for creating SSLForwardProxy instances."""
-
-    class Meta:
-        model = SSLForwardProxy
-
-    auto_include_altname = False
-    block_client_cert = False
-    block_expired_certificate = False
-    block_timeout_cert = False
-    block_tls13_downgrade_no_resource = False
-    block_unknown_cert = False
-    block_unsupported_cipher = False
-    block_unsupported_version = False
-    block_untrusted_issuer = False
-    restrict_cert_exts = False
-    strip_alpn = False
-
-    @classmethod
-    def with_all_blocks_enabled(cls, **kwargs):
-        """Create an instance with all blocking options enabled."""
-        return cls(
-            block_client_cert=True,
-            block_expired_certificate=True,
-            block_timeout_cert=True,
-            block_tls13_downgrade_no_resource=True,
-            block_unknown_cert=True,
-            block_unsupported_cipher=True,
-            block_unsupported_version=True,
-            block_untrusted_issuer=True,
-            **kwargs,
-        )
-
-
-class SSLInboundProxyFactory(factory.Factory):
-    """Factory for creating SSLInboundProxy instances."""
-
-    class Meta:
-        model = SSLInboundProxy
-
-    block_if_hsm_unavailable = False
-    block_if_no_resource = False
-    block_unsupported_cipher = False
-    block_unsupported_version = False
-
-    @classmethod
-    def with_all_blocks_enabled(cls, **kwargs):
-        """Create an instance with all blocking options enabled."""
-        return cls(
-            block_if_hsm_unavailable=True,
-            block_if_no_resource=True,
-            block_unsupported_cipher=True,
-            block_unsupported_version=True,
-            **kwargs,
-        )
-
-
-class SSLNoProxyFactory(factory.Factory):
-    """Factory for creating SSLNoProxy instances."""
-
-    class Meta:
-        model = SSLNoProxy
-
-    block_expired_certificate = False
-    block_untrusted_issuer = False
-
-    @classmethod
-    def with_all_blocks_enabled(cls, **kwargs):
-        """Create an instance with all blocking options enabled."""
-        return cls(
-            block_expired_certificate=True,
-            block_untrusted_issuer=True,
-            **kwargs,
-        )
-
-
-# SDK tests against SCM API
-class DecryptionProfileCreateApiFactory(factory.Factory):
-    """Factory for creating DecryptionProfileCreateModel instances."""
-
-    class Meta:
-        model = DecryptionProfileCreateModel
-
-    name = factory.Sequence(lambda n: f"decryption_profile_{n}")
-    folder = "Texas"
-    ssl_protocol_settings = factory.SubFactory(SSLProtocolSettingsFactory)
-    ssl_forward_proxy = factory.SubFactory(SSLForwardProxyFactory)
-    ssl_inbound_proxy = factory.SubFactory(SSLInboundProxyFactory)
-    ssl_no_proxy = factory.SubFactory(SSLNoProxyFactory)
-
-    @classmethod
-    def with_snippet(cls, snippet: str = "TestSnippet", **kwargs):
-        """Create an instance with snippet container."""
-        return cls(folder=None, snippet=snippet, device=None, **kwargs)
-
-    @classmethod
-    def with_device(cls, device: str = "TestDevice", **kwargs):
-        """Create an instance with device container."""
-        return cls(folder=None, snippet=None, device=device, **kwargs)
-
-    @classmethod
-    def with_custom_ssl_settings(
-        cls,
-        min_ver: SSLVersion = SSLVersion.tls1_0,
-        max_ver: SSLVersion = SSLVersion.tls1_2,
-        **kwargs,
-    ):
-        """Create an instance with custom SSL protocol settings."""
-        return cls(
-            ssl_protocol_settings=SSLProtocolSettingsFactory.with_versions(
-                min_ver=min_ver, max_ver=max_ver
-            ),
-            **kwargs,
-        )
-
-
-class DecryptionProfileUpdateApiFactory(factory.Factory):
-    """Factory for creating DecryptionProfileUpdateModel instances."""
-
-    class Meta:
-        model = DecryptionProfileUpdateModel
-
-    id = factory.LazyFunction(lambda: str(uuid.uuid4()))
-    name = factory.Sequence(lambda n: f"decryption_profile_{n}")
-    ssl_protocol_settings = factory.SubFactory(SSLProtocolSettingsFactory)
-    ssl_forward_proxy = factory.SubFactory(SSLForwardProxyFactory)
-    ssl_inbound_proxy = factory.SubFactory(SSLInboundProxyFactory)
-    ssl_no_proxy = factory.SubFactory(SSLNoProxyFactory)
-
-    @classmethod
-    def with_updated_ssl_settings(cls, **kwargs):
-        """Create an instance with updated SSL protocol settings."""
-        return cls(
-            ssl_protocol_settings=SSLProtocolSettingsFactory(
-                min_version=SSLVersion.tls1_1,
-                max_version=SSLVersion.tls1_3,
-            ),
-            **kwargs,
-        )
-
-
-class DecryptionProfileResponseFactory(factory.Factory):
-    """Factory for creating DecryptionProfileResponseModel instances."""
-
-    class Meta:
-        model = DecryptionProfileResponseModel
-
-    id = factory.LazyFunction(lambda: str(uuid.uuid4()))
-    name = factory.Sequence(lambda n: f"decryption_profile_{n}")
-    folder = "Texas"
-    ssl_protocol_settings = factory.SubFactory(SSLProtocolSettingsFactory)
-    ssl_forward_proxy = factory.SubFactory(SSLForwardProxyFactory)
-    ssl_inbound_proxy = factory.SubFactory(SSLInboundProxyFactory)
-    ssl_no_proxy = factory.SubFactory(SSLNoProxyFactory)
-
-    @classmethod
-    def with_snippet(cls, snippet: str = "TestSnippet", **kwargs):
-        """Create an instance with snippet container."""
-        return cls(folder=None, snippet=snippet, device=None, **kwargs)
-
-    @classmethod
-    def with_device(cls, device: str = "TestDevice", **kwargs):
-        """Create an instance with device container."""
-        return cls(folder=None, snippet=None, device=device, **kwargs)
-
-    @classmethod
-    def from_request(cls, request_model: DecryptionProfileCreateModel, **kwargs):
-        """Create a response model based on a request model."""
-        data = request_model.model_dump()
-        data["id"] = str(uuid.uuid4())
-        data.update(kwargs)
-        return cls(**data)
-
-
-# Pydantic modeling tests
-class DecryptionProfileCreateModelFactory(factory.DictFactory):
-    """Factory for creating data dicts for DecryptionProfileCreateModel validation testing."""
-
-    name = factory.Sequence(lambda n: f"decryption_profile_{n}")
-    folder = "Texas"
-
-    @classmethod
-    def build_valid(cls):
-        """Return a valid data dict with all expected attributes."""
-        return cls(
-            name="TestProfile",
-            folder="Texas",
-            ssl_protocol_settings={
-                "min_version": "tls1-0",
-                "max_version": "tls1-2",
-            },
-        )
-
-    @classmethod
-    def build_with_invalid_name(cls):
-        """Return a data dict with invalid name pattern."""
-        return cls(
-            name="@invalid-name#",
-            folder="Texas",
-        )
-
-    @classmethod
-    def build_with_invalid_ssl_versions(cls):
-        """Return a data dict with invalid SSL version configuration."""
-        return cls(
-            name="TestProfile",
-            folder="Texas",
-            ssl_protocol_settings={
-                "min_version": "tls1-2",
-                "max_version": "tls1-0",  # Invalid: max < min
-            },
-        )
-
-    @classmethod
-    def build_with_multiple_containers(cls):
-        """Return a data dict with multiple containers."""
-        return cls(
-            name="TestProfile",
-            folder="Texas",
-            snippet="TestSnippet",
-        )
-
-    @classmethod
-    def build_with_no_container(cls):
-        """Return a data dict without any container."""
-        return cls(
-            name="TestProfile",
-            folder=None,
-        )
-
-
-class DecryptionProfileUpdateModelFactory(factory.DictFactory):
-    """Factory for creating data dicts for DecryptionProfileUpdateModel validation testing."""
-
-    id = "123e4567-e89b-12d3-a456-426655440000"
-    name = factory.Sequence(lambda n: f"decryption_profile_{n}")
-
-    @classmethod
-    def build_valid(cls):
-        """Return a valid data dict for updating a decryption profile."""
-        return cls(
-            id="123e4567-e89b-12d3-a456-426655440000",
-            name="UpdatedProfile",
-            ssl_protocol_settings={
-                "min_version": "tls1-1",
-                "max_version": "tls1-3",
-            },
-        )
-
-    @classmethod
-    def build_with_invalid_fields(cls):
-        """Return a data dict with multiple invalid fields."""
-        return cls(
-            id="invalid-uuid",
-            name="@invalid-name",
-            ssl_protocol_settings={
-                "min_version": "invalid-version",
-                "max_version": "also-invalid",
-            },
-        )
-
-    @classmethod
-    def build_minimal_update(cls):
-        """Return a data dict with minimal valid update fields."""
-        return cls(
-            id="123e4567-e89b-12d3-a456-426655440000",
-            ssl_forward_proxy={"block_client_cert": True},
-        )
-
 
 # ----------------------------------------------------------------------------
 # Security Rule object factories.
